==> code/04_colors.R <==
colM = c(rgb(0,.45,.7), rgb(0,.45,.8), rgb(0,.45,.6))
colF = c(rgb(0.3,.9,.9),rgb(0,.65,.65),rgb(0,.55,.55),rgb(0,.75,.75))
 
colAh = c(rgb(1,.8,.2),rgb(1,.88,.2), rgb(1,.72,.2))
colAm = c(rgb(.9,.5,0),rgb(.9,.55,0),rgb(.9,.3,0))
colAl = c(rgb(.75,0,0),rgb(.9,0,0),  rgb(.55,0,0))
